+++ lib/braintree/transaction.rb
@@ -121,6 +121,8 @@ module Braintree
     attr_reader :google_pay_details
     attr_reader :graphql_id
     attr_reader :id
+    attr_reader :installment_count
+    attr_reader :installments
     attr_reader :local_payment_details
     attr_reader :merchant_account_id
     attr_reader :network_response_code                  # Response code from the card network
@@ -143,6 +145,7 @@ module Braintree
     attr_reader :recurring
     attr_reader :refund_ids
     attr_reader :refunded_transaction_id
+    attr_reader :refunded_installments
     attr_reader :retrieval_reference_number
     attr_reader :risk_data
     attr_reader :samsung_pay_card_details
@@ -306,6 +309,9 @@ module Braintree
       @visa_checkout_card_details = VisaCheckoutCardDetails.new(attributes[:visa_checkout_card])
       @samsung_pay_card_details = SamsungPayCardDetails.new(attributes[:samsung_pay_card])
       authorization_adjustments.map! { |attrs| AuthorizationAdjustment._new(attrs) } if authorization_adjustments
+
+      installments.map! { |attrs| Installment.new(attrs) } if installments
+      refunded_installments.map! { |attrs| Installment.new(attrs) } if refunded_installments
     end
 
     def inspect # :nodoc: